scripts/requirements.txt
pygbif
biopython==1.*
jinja2
filelock
pycountry
fsspec
geopandas
pyyaml
markdown2